+++ file.doc
@@ -1,0 +1,112 @@
+<? Doc Format>
+[
+1 фамилия: Иванов, имя: Владмир
+ дата рождения: 1925-10-04, дата смерти: 1987-04-01
+ пол мужской: Male
+ отец не определён 
+ мать не определёна 
+ дети: 
+ сын :  id: 9 фамилия: Иванов имя: Иван дата рождения: 1950-03-25
+, 
+9 фамилия: Иванов, имя: Иван
+ дата рождения: 1950-03-25 возраст: 73
+ пол мужской: Male
+ отец id: 1 фамилия: Иванов, имя:  Владмир, год рождения: 1925-10-04
+ мать id: 3 фамилия: Иванова, имя:  Екатерина, год рождения: 1927-07-23
+ дети: 
+ сын :  id: 13 фамилия: Иванов имя: Антон дата рождения: 1980-05-28
+, 
+13 фамилия: Иванов, имя: Антон
+ дата рождения: 1980-05-28 возраст: 42
+ пол мужской: Male
+ отец id: 9 фамилия: Иванов, имя:  Иван, год рождения: 1950-03-25
+ мать id: 11 фамилия: Михайлова, имя:  Мария, год рождения: 1953-12-30
+ детей нет
+, 
+3 фамилия: Иванова, имя: Екатерина
+ дата рождения: 1927-07-23, дата смерти: 1984-10-02
+ женский пол:Female
+ отец не определён 
+ мать не определёна 
+ дети: 
+ сын :  id: 9 фамилия: Иванов имя: Иван дата рождения: 1950-03-25
+, 
+7 фамилия: Матвеев, имя: Антон
+ дата рождения: 1923-05-17, дата смерти: 1995-10-09
+ пол мужской: Male
+ отец не определён 
+ мать не определёна 
+ дети: 
+ дочь:  id: 12 фамилия: Матвеева имя: Елена дата рождения: 1950-09-28
+, 
+8 фамилия: Матвеева, имя: Екатерина
+ дата рождения: 1928-09-28, дата смерти: 1995-03-25
+ женский пол:Female
+ отец не определён 
+ мать не определёна 
+ дети: 
+ дочь:  id: 12 фамилия: Матвеева имя: Елена дата рождения: 1950-09-28
+, 
+12 фамилия: Матвеева, имя: Елена
+ дата рождения: 1950-09-28 возраст: 72
+ женский пол:Female
+ отец id: 7 фамилия: Матвеев, имя:  Антон, год рождения: 1923-05-17
+ мать id: 8 фамилия: Матвеева, имя:  Екатерина, год рождения: 1928-09-28
+ дети: 
+ сын :  id: 14 фамилия: Петров имя: Кирилл дата рождения: 1983-11-20
+, 
+5 фамилия: Михайлов, имя: Роман
+ дата рождения: 1923-10-05, дата смерти: 1990-02-04
+ пол мужской: Male
+ отец не определён 
+ мать не определёна 
+ дети: 
+ дочь:  id: 11 фамилия: Михайлова имя: Мария дата рождения: 1953-12-30
+, 
+6 фамилия: Михайлова, имя: Екатерина
+ дата рождения: 1925-05-08, дата смерти: 1993-10-04
+ женский пол:Female
+ отец не определён 
+ мать не определёна 
+ дети: 
+ дочь:  id: 11 фамилия: Михайлова имя: Мария дата рождения: 1953-12-30
+, 
+11 фамилия: Михайлова, имя: Мария
+ дата рождения: 1953-12-30 возраст: 69
+ женский пол:Female
+ отец id: 5 фамилия: Михайлов, имя:  Роман, год рождения: 1923-10-05
+ мать id: 6 фамилия: Михайлова, имя:  Екатерина, год рождения: 1925-05-08
+ дети: 
+ сын :  id: 13 фамилия: Иванов имя: Антон дата рождения: 1980-05-28
+, 
+2 фамилия: Петров, имя: Николай
+ дата рождения: 1928-05-10, дата смерти: 1980-03-17
+ пол мужской: Male
+ отец не определён 
+ мать не определёна 
+ дети: 
+ сын :  id: 10 фамилия: Петров имя: Пётр дата рождения: 1948-11-11
+, 
+10 фамилия: Петров, имя: Пётр
+ дата рождения: 1948-11-11 возраст: 74
+ пол мужской: Male
+ отец id: 2 фамилия: Петров, имя:  Николай, год рождения: 1928-05-10
+ мать id: 4 фамилия: Петрова, имя:  Антонина, год рождения: 1928-06-11
+ дети: 
+ сын :  id: 14 фамилия: Петров имя: Кирилл дата рождения: 1983-11-20
+, 
+14 фамилия: Петров, имя: Кирилл
+ дата рождения: 1983-11-20 возраст: 39
+ пол мужской: Male
+ отец id: 10 фамилия: Петров, имя:  Пётр, год рождения: 1948-11-11
+ мать id: 12 фамилия: Матвеева, имя:  Елена, год рождения: 1950-09-28
+ детей нет
+, 
+4 фамилия: Петрова, имя: Антонина
+ дата рождения: 1928-06-11, дата смерти: 1990-01-20
+ женский пол:Female
+ отец не определён 
+ мать не определёна 
+ дети: 
+ сын :  id: 10 фамилия: Петров имя: Пётр дата рождения: 1948-11-11
+]
