+++ src/model/fileWork/Format/ReadDoc/interfaceRead.java
@@ -1,0 +1,8 @@
+package model.fileWork.Format.ReadDoc;
+
+import java.io.IOException;
+
+public interface interfaceRead<T>
+{
+    String read(T t) throws IOException, ClassNotFoundException;
+}
