+++ src/model/fileWork/Format/SaveDoc/InterfaceSave.java
@@ -1,0 +1,8 @@
+package model.fileWork.Format.SaveDoc;
+
+import java.io.IOException;
+
+public interface InterfaceSave<T>
+{
+    void write(T t) throws IOException, ClassNotFoundException;
+}
