+++ src/model/fileWork/InterfaceSave.java
@@ -1,0 +1,5 @@
+package model.fileWork;
+
+public class InterfaceSave {
+
+}
