+++ src/view/Starts/interfaceStart.java
@@ -1,0 +1,6 @@
+package view.Starts;
+
+public interface interfaceStart 
+{
+   public int start();
+}
