+++ src/view/tablo/interfaceTablo.java
@@ -1,0 +1,6 @@
+package view.tablo;
+
+public interface interfaceTablo 
+{
+    public void tablo();
+}
